+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ProofNetworkJdf Changelog
 #### Features
  * Component::length renamed to Component::depth
  * Short version of Component Dimensions attribute (with 2 parts) is now supported
+ * Partitioned Layouts are now supported
 
 #### Bug Fixing
  * --